requirements.txt
aiogram==3.15.0
boto3==1.35.99
aiohttp==3.10.11
python-dotenv==1.0.1
